pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundation-model-eval-harness"
version = "0.1.0"
description = "Config-driven foundation model evaluation harness for biomedical downstream tasks"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "langchain>=0.3.27",
  "langchain-community>=0.3.7",
  "langgraph>=0.2.60",
  "transformers>=4.46.0",
  "datasets>=3.1.0",
  "evaluate>=0.4.3",
  "torch>=2.4.0",
  "scikit-learn>=1.5.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pydantic>=2.9.0",
  "duckdb>=1.1.0",
  "mlflow>=2.17.0",
  "hydra-core>=1.3.2",
  "typer>=0.12.5",
  "rich>=13.9.4",
  "jinja2>=3.1.4",
  "matplotlib>=3.9.2",
  "sentence-transformers>=3.2.1",
  "faiss-cpu>=1.9.0",
  "pyarrow>=18.0.0",
  "streamlit>=1.40.0",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pre-commit>=4.0.1",
  "ruff>=0.7.3",
  "black>=24.10.0",
  "isort>=5.13.2",
  "mypy>=1.13.0",
  "types-PyYAML>=6.0.12.20240917",
]

[project.scripts]
fmeh = "fmeh.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
